more flexible (write') function; speedup
We use the improved (write') function to modify the EDB file as well as
the EDU file. The new function takes two delimiters; one is the regex
version and the other is the literal version of the same thing.

The regexes on the EDB file took about a second or more because of the
many regexes involved; by using the (write') function, we get
near-instant results.