For each building type (walls, barracks, stables, ports, etc.), let all levels of that building recruit the same units. However, we give the more advanced buildings bonuses to base unit number, recruitment points gained per turn, maximum recruitment points available, and (most importantly) the starting experience.
This way, you don't have to wait 1/2 or 2/3 of the game to get a chance
to play with elite units. And plus, you can finally see a sizeable
number of units with silver/gold chevrons (experience) without resorting
to moving around old, stale stacks of units that take many beatings to
gain experience.

Also, as an added bonus, you can retrain most units at far-away
settlements, as long as they have at least the very lowest level of the
building in question.