m2tw: re-disable princess bowing animation
This undoes the change introduced in
e9067122ab02457036a266b3982dc5383f4b3106.

This way, other mods can just use the packed animation files only for
this feature (since other mods most likely have princesses in them).

Also, note that we had not been tracking the changes to the animation
settings in the actual binary diffs. Just taking note of this for the
record.