rtw: use original descr_strat.txt
We use the original file; the only reason for using the modified version
was because this allowed existing RTW players who already modified their
original descr_strat.txt file (by virtue of playing and beating the
imperial campaign, thus unlocking all 8 of the unlockable factions) to
run quicksilver. But, we don't need to do this, because:
        - quicksilver is meant to be used on a pristine vanilla RTW
          folder, and
        - this allows players to quickly install Steam + RTW, and get
          started with quicksilver right away.

In the event that the player does not already have the 8 factions
unlocked, he can successfully pass the SHA checksum check by cut/pasting
the 8 unlockable factions to the bottom of the playable factions list
(just below the Roman factions).