animations: use caliban's unpacked anim data
We want to use vanilla steam animation files, but we temporarily use the
old unpacked files as the reference. The animations seem to work, at
least in the beginning of the campaign. This is a temporary fix. Should
the mod crash in the campaign map, the solution should be to simply
delete the "data/animations" folder in the mod.

We define a new "anims_binary_dir" folder because we cannot use steam's
vanilla installation folder, as the binary diffs we have can only be
applied to the old SHA-1-sum-matching files from
caliban's/modding-community's binary packfiles. (Grep "Caliban" for more
info).

The SHA checksum changes to the sound files are necessary because we are
now using the vanilla steam sound files (and not the old non-steam
files). The same goes for the "map_FE" to "map_fe" filename change.

These changes were necessary to allow qs to generate the mods as before.

Conflicts:
	src/Check.hs
	src/Option.hs
	src/qs.hs